Add lens for /etc/system (Solaris)
File is used to configure the kernel, module loading and parameters. The
system(4) man page gives a good overview of the commands and a sample file:
http://docs.sun.com/app/docs/doc/816-5174/system-4?l=en&n=1&a=view

It supports moddir, rootdev, rootfs, exclude, include, forceload and set.
The set tree is broken down quite a lot as the operator can vary (equals,
bitwise AND and OR).