* pathx.c (ns_filter): create new nodeset instead of in-place filtering
In-place filtering of a large number of potential nodes down to a single
node would be very inefficient due to use of ns_remove calling memmove
to remove single elements from the list.

memmove would be continually called, shifting the contents of the
nodeset one element at a time, causing many unnecessary copies.

Time taken to perform aug_set and aug_get from 1-5000 with
/test/service[1..5000] is now:

    testPerfPredicate = 52061ms

(A reduction of 21.9% over the previous commit.)