* pathx.c (ns_add): only check for node in nodeset on union
ns_add now does not check if a node is already in the nodeset before
adding it, and eval_union now calls ns_add_dedupe as it's the only code
path needing this behaviour.

This improves performance drastically from ns_from_locpath and ns_filter
which start with empty nodesets.  In a benchmark calling:

    aug_set("/test/service[1]", "foo")  from 1..750
    aug_get("/test/service[1]")         from 1..750

the ns_from_locpath function represents 36.44% of the test instruction
count, in which  the ns_add function was called 858,585 times,
accounting for 90.43% of the instructions in ns_from_locpath.

Of this, 98.09% was spent in the initial loop checking for the node's
presence in the nodeset.  Only 0.63% was spent on memory reallocations.

After this patch, ns_from_locpath represents only 6.08% of the whole
benchmark and ns_add only 0.93% of ns_from_locpath's entire instruction
count.

Performance of testPerfPredicate is now:

    testPerfPredicate = 1421ms

(A reduction of 97.3% over the previous commit.)